(PC-14474) refactor(Booking): add tests for QrCode, NoTicket, EmailSent, TicketBody and DefaultBody
