+++ src/features/bookings/components/BookingDetailsTicketContent.test.tsx
@@ -1,0 +1,82 @@
+import React from 'react'
+
+import { BookingDetailsTicketContent } from 'features/bookings/components/BookingDetailsTicketContent'
+import { render } from 'tests/utils'
+
+import { bookingsSnap } from '../api/bookingsSnap'
+
+describe('BookingDetailsTicketContent', () => {
+  describe('activationCodeFeatureEnabled & token', () => {
+    it('should display the booking token when activationCodeFeatureEnabled is false', () => {
+      const booking = { ...bookingsSnap.ongoing_bookings[0], activationCode: { code: 'someCode' } }
+      const { getByText, queryByText } = render(
+        <BookingDetailsTicketContent booking={booking} activationCodeFeatureEnabled={false} />
+      )
+
+      getByText(booking.token)
+      expect(queryByText(booking.activationCode.code)).toBeFalsy()
+    })
+    it('should display the booking activation code when activationCodeFeatureEnabled is true', () => {
+      const booking = { ...bookingsSnap.ongoing_bookings[0], activationCode: { code: 'someCode' } }
+      const { getByText, queryByText } = render(
+        <BookingDetailsTicketContent booking={booking} activationCodeFeatureEnabled={true} />
+      )
+
+      getByText(booking.activationCode.code)
+      expect(queryByText(booking.token)).toBeFalsy()
+    })
+  })
+})
+
+// describe('QR code display', () => {
+//   const initialBooking = {
+//     ...bookingsSnap.ongoing_bookings[0],
+//     stock: {
+//       ...bookingsSnap.ongoing_bookings[0].stock,
+//       offer: {
+//         ...bookingsSnap.ongoing_bookings[0].stock.offer,
+//         isDigital: false,
+//       },
+//     },
+//   }
+
+//   // TODO => A déplacer dans ticketBodyMessage
+//   it('should display information like what the event is today', () => {
+//     const booking = {
+//       ...initialBooking,
+//       confirmationDate: formatISO(addDays(new Date(), -2)).slice(0, -1),
+//       stock: {
+//         ...initialBooking.stock,
+//         beginningDatetime: formatISO(new Date()).slice(0, -1),
+//         offer: {
+//           ...initialBooking.stock.offer,
+//           subcategoryId: SubcategoryIdEnum.FESTIVAL_MUSIQUE,
+//           withdrawalType: WithdrawalTypeEnum.by_email,
+//           withdrawalDelay: 0,
+//         },
+//       },
+//     }
+
+//     const { queryByTestId } = render(<BookingDetailsTicketContent booking={booking} />)
+
+//     expect(queryByTestId('withdrawal-info-email-msg')?.children[0]).toContain(`aujourd'hui`)
+//   })
+
+//   it('should display no ticket withdrawal wording and circled check icon', () => {
+//     const booking = {
+//       ...initialBooking,
+//       stock: {
+//         ...initialBooking.stock,
+//         offer: {
+//           ...initialBooking.stock.offer,
+//           subcategoryId: SubcategoryIdEnum.FESTIVAL_MUSIQUE,
+//           withdrawalType: WithdrawalTypeEnum.no_ticket,
+//           withdrawalDelay: 0,
+//         },
+//       },
+//     }
+
+//     const { getByTestId } = render(<BookingDetailsTicketContent booking={booking} />)
+//     getByTestId('withdrawal-info-no-ticket')
+//   })
+// })
