+++ src/features/bookings/components/TicketBody/EmailSent/EmailSent.web.test.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import waitForExpect from 'wait-for-expect'
+
+import { EmailSent } from 'features/bookings/components/TicketBody/EmailSent/EmailSent'
+import { render } from 'tests/utils/web'
+
+const offerDate = new Date(2022, 11, 17)
+
+describe('<EmailSent/>', () => {
+  it('should not display the button "Consulter mes e-mails"', async () => {
+    const { queryByTestId } = render(<EmailSent offerDate={offerDate} />)
+    await waitForExpect(() => {
+      expect(queryByTestId('withdrawal-info-email-btn')).toBeFalsy()
+    })
+  })
+})
